+++ tests/Fixtures/Author.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Coyote6\LaravelBase\Tests\Fixtures;
+
+class Author
+{
+    //
+}